predicted-gh-rate-limit-exhaustion: set duration to 60s
On about May 11 1AM PDT, we got 3 sub-1-minute alerts for istio-testing
for this metric:

- May 11, 2022 at 8:00AM UTC (28 sec ago)
       Alert firing lasted
       27 secs
- May 11, 2022 at 8:01AM UTC (29 sec ago)
       Alert firing lasted
       28 secs
- May 11, 2022 at 8:02AM UTC (32 sec ago)
       Alert firing lasted
       31 secs

This was due to a sudden drop in tokens over this time period, which was
quickly recovered. It's not clear why istio-testing experiences sudden
drops for this metric, triggering this alert, but from an administrative
point of view it probably doesn't help that we have multiple alerts fire
for less than a minute.

This change sets the duration field to 60s (it can only be in multiples
of 60s [1]), which is still a conservative figure.

[1]: https://registry.terraform.io/providers/hashicorp/google/latest/docs/resources/monitoring_alert_policy#duration