increase CS e2e memory request to 80Gi
The large-job-pool on our prow build cluster uses c2-standard-30 nodes,
which come with 30 vCPUs and 120GB memory. These jobs request 29 vCPUs
and 40Gi memory, meaning each job can only run on a single node but
underutilizes the memory on that node. This increases the memory request
to improve memory utilization on the node. This will hopefully reduce
test flakiness.