Adjust the CPU/mem resources
The peak CPU usage of the presubmit mono/multi jobs is about 10 CPUs;
The peak mem usage of the presubmit mono/multi jobs is about 16 GiBs.

This PR sets the CPU/mem requests to 17CPUs/30GiBs, so that we can support running
three presubmit jobs on a single node concurrently.