freeze prow deployments
The `ci-oss-test-infra-autobump-prow` job creates the bump PR. Here we
delete the job (`ci-oss-test-infra-autobump-prow-for-auto-deploy`) that
votes on these PRs to skip review, making them merge automatically. When
we are back from holidays we can restore this job so that
Prow (oss.gprow.dev) is updated automatically again.

For examples of autobump PRs, see
https://github.com/GoogleCloudPlatform/oss-test-infra/pulls?q=is%3Apr+author%3Agoogle-oss-robot+is%3Aclosed.