Extend 7-autopilot-rapid-latest timeout to th
NOTE: this change is to be reverted.

The CI for this specific has been failing with inconsistent failures, with some of them being 4h timeouts.

Extending the timeout temporarily to 5h to see if there could be a successful run. In the long term we should optimize the test to take less time.