Add a CI job for ConfigSync stress test on Autopilot rapid-latest
It also changes the stress test to run on the Standard rapid-latest channel
instead of the regular channel to match the Autopilot one.

We choose the rapid-latest channel because we observed that the latest
GKE version often time consumes more resources, so we use that to detect
stress test failures.