Use a separate large job pool for the release CI
The kind job on both the main branch and the release branch failed
because of api timeout.
After checking the node status, we are seeing MemoryPressure, DiskPressure,
and PIDPressure issues on the node.

Hence, this commit moves the kind job on the release branch to a
separate large job pool.

b/299570441