add gke-gcloud-auth-plugin kubeconfig file for the oss-prow
The oss-prow cluster had its CA root certificate (public key) rotated
earlier this week because it was going to expire next month. However
this means that our Prow components need to update their kubeconfigs to
schedule trusted jobs (e.g., Prow Controller Manager needs to talk to
the K8s API server on the oss-prow cluster), because they still use the
old kubeconfig for the oss-prow cluster with the old CA root
certificate.

Check in the minimal kubeconfig for access into the oss-prow cluster.
This minimal configuration was created as follows:

First we ran

    gcloud --project=oss-prow container clusters get-credentials prow --region us-west1-a

which fetched the correct cluster IP (which was rotated as part of the
CA rotation) and CA certificate (public key), but which uses the new
gke-gcloud-auth-plugin to obtain client credentials. The
kubeconfigs.yaml content was the result of running

    kubectl config view --context=oss -oyaml >kubeconfigs.yaml

followed by removing all other non-oss-prow cluster entries. The
"certificate-authority-data" field was not exported with the above
command (it just had "DATA+OMITTED"), so we had to look it up directly in
the raw "~/.kube/config" file manually. This file was tested with

    kubectl --kubeconfig ./kubeconfigs.yaml get po

to check that it worked with the oss-prow cluster.

As part of this change, also create a Makefile.gcloud.mk rule to update this
kubeconfigs.yaml file going forward (we might want to add in kubeconfigs
for other clusters here in the future). The rule here is essentially
copied from the update-config rule in the main Makefile. Note that we
must first run

    kubectl create configmap kubeconfigs --from-file=kubeconfigs.yaml=kubeconfigs.yaml

in order to first create the configmap in the cluster, after which the

    ... | kubectl replace configmap kubeconfigs -f -

will work (otherwise this part will fail because the "kubeconfigs"
configmap doesn't exist yet).

A subsequent change will mount the configmap into our components (such
as Prow Controller Manager) that need it.