fix kubeconfig for test-infra-trusted cluster
The update to the kubeconfigs.yaml file has already been applied to the
oss-prow cluster, and is a NOP.

Move the kubeconfigs.yaml file into a subdirectory under cluster/. This
keeps it closer to the other cluster-specific configs. Putting it under
a subdirectory makes

    kubectl apply -f ./cluster/

in the "deploy" target skip over it.