Update kpt-config-sync presubmit image
Use the go-canary image variant, because it includes Go 1.21, which
is required to run `go mod tidy` in presubmits. Previously we pinned
to the 1.26 variant, to ensure a specific version of k8s was used.
The go-canary variant is pinned to the `stable` version of k8s, which
today is v1.28.2 (per https://cdn.dl.k8s.io/release/stable.txt).