Revamp guest-test-infra presubmit checks
We are missing a lot of gobuild/gocheck/gotest coverage because these containers are scoped to only subdirectories which belong to the same module as the main directory. This change will add each module as its own presubmit directory to ensure coverage.

citpresubmit still handles building and testing because of the way we handle test code in those directories.