add manual k8s CronJob to regularly restart KES
The problem with kubernetes-external-secrets (KES) is that it can start
failing, but where the process does not exit. So from Kubernetes' point
of view, it doesn't know that it should restart KES.

Inform our Prow instance that we want to restart KES regularly, with a
CronJob. This has the advantage that it will continue to run even if
there are any issues with Prow itself.

Tested locally against my local Prow instance.