kes-kicker: target 'default' namespace
In #2125, we created various resources (ServiceAccount, Role,
RoleBinding, CronJob) all with the name "kes-kicker". Looking at the
cluster after these configurations have been applied, they all appear to
be in the "test-pods" namespace.

Consequently, we get an error like this in the logs from the kes-kicker
pod when it gets scheduled to satisfy the CronJob definition:

    Error from server (NotFound): deployments.apps "kubernetes-external-secrets" not found

This appears to be because the kubectl command inside kes-kicker is
operating inside the test-pods namespace.

Recreate the same resources in the default namespace. But also, give the
kubectl invocation an namespace because it's more strict that way and it
can't hurt.