Change the default Git provider to CSR
The kind test covers testing with the local test-git-server. This commit
changes other CI jobs to test with the Cloud Source Repository so that
we don't need a separate job for Git + WI tests.

It also removes the test regex for those one-off jobs because
https://github.com/GoogleContainerTools/kpt-config-sync/pull/1160
changes the e2e test to run these tests if and only if the flag is
specified.

b/325119647