Run performance periodics on a longer, lower interval
This will make more efficient use of exisiting quota. The images in cit-performance-periodics change very infrequently, the images in ubuntu-daily-cit-performance change more frequently so have a slightly higher interval. Expecting runtimes of 8.4 - 19.7 hours (with an average of 11.4 hours), which gives a four hour buffer before timeout. parallel_count=1 is more efficient but doubles these runtimes and running a prow job for longer than 24 hours requires some extra configuration, so sticking with this for now.