enable MTU workaround for dind presets
This fixes an issue where the dind container is on a host with a MTU
lower than the typical default of 1500, resulting in a mismatch.

Upstream maintainers are considering making this the default behavior of
the script, but for now this is a quick workaround.

See: https://github.com/kubernetes/test-infra/blob/94f76aedb29b50131fe5dc5e131adcd23f6d4f22/images/bootstrap/runner.sh#L87