bump timeout for config sync autopilot tests
It appears that new versions of autopilot (currently only in rapid
channel) scale up slower, and as a result the tests are occasionally
timing out. The average runtime is very close to the current 2h timeout,
causing it to sometimes pass and sometimes time out.

A 4h timeout matches the current timeout set on the prow job config.