Trigger TestGrid jobs on main, not master, branch.
We switched from a master branch to a main branch a while back in the
GCP/testgrid repo, but apparently neglected to update the related jobs.
(I suspect this got missed during the branch change due to other
priorities / team changes around that time).

Jobs may need fixes after this is submitted, but this should at least
get them up and running again.