Revert oracle-toolkit Ansible runner version
When run with ansible 2.11, tests are failing with the error

```
ERROR! couldn't resolve module/action 'mount'. This often indicates a
misspelling, missing collection, or incorrect module path.

The error appears to be in
'/home/prow/go/src/github.com/google/oracle-toolkit/roles/brute-ora-cleanup/tasks/main.yml':
line 231, column 3, but may be elsewhere in the file depending on the exact syntax problem.
```

Testing with 2.12 I see the same error, but in 2.9, the error goes away.
So for now let's revert to 2.9 to unblock tests.

Successful output (internal):
https://paste.googleplex.com/4939247786917888?raw