Switch Rocky accelerator tests to use test images and add Rocky 10
The Rocky accelerator test are currently pointing to the production images
in rocky-linux-accelerator-cloud, but these are the only tests we currently
have for mRDMA, so let's point them to the test images in gce-ciq-images.

The production images are guaranteed to be identical to the test images at
the time they are released, so this gives us a way to ensure mRDMA is
working as expected *before* the images go to production.

This also allows us to test the Rocky 10 Accelerator images, which have not
yet been pushed to production.

Signed-off-by: Jonathan Dieter <jdieter@ciq.com>